Implement the ANSI class
Gary took advantage of the fact that Ruby uses the `!` character as a
convention to let the user know this changes state. There is no such
affordance in Rust and the `!` is not allowed in function names. I
thought about prefixing the _pure_ functions with an underscore, but
that tells Rust to not throw a warning if that function is not used. In
the end, I decided the pure functions do not need to be part of the
`ANSI` struct and moved them out into `mod ansi` as normal functions.