Move TTY to its own module
As I understand it, the reason Gary had all of selecta in one file is
bceause that file was acting as the single executable required to run
selecta. Having a build step is not something Gary wanted to do, so this
was the strategy that made it easy for people to install.

In Rust, we have to compile. We are better off then separating our code
into separate files to take advantage of the module system.